backend/requirements.txt
fastapi==0.115.0
uvicorn==0.30.0
python-multipart==0.0.9
sentence-transformers==3.0.0
faiss-cpu==1.8.0
fpdf2==2.8.1
pydantic==2.9.0
python-dotenv==1.0.1
google-generativeai==0.8.0
fuzzywuzzy==0.18.0
python-Levenshtein==0.25.1
numpy==1.26.4
